Successful Merge

+++ WhichIsBigger/Classes/Data Model/WIBGameOption.h
@@ -14,7 +14,7 @@
 - (id)initWithItem:(WIBGameItem *)item;
 
 @property (nonatomic, strong) WIBGameItem *item;
-@property (nonatomic) double multiplier;
+@property (nonatomic) int multiplier;
 @property (nonatomic, readonly) NSNumber *total;
 
 @end
